Added toast to application, added Helper class in web for Toast notification parsing, fixed bug in CreateCategory.razor.cs where if you leave the modal and open it again the last categoryName would still be written in field, fixed a similar bug in UpdateCategory.razor where changes would persist after closing modal, handled null cases

+++ Web/Components/Pages/Categories/CategoryModals/DeleteCategory.razor.cs
@@ -7,20 +7,24 @@ namespace Web.Components.Pages.Categories.CategoryModals;
 
 public partial class DeleteCategory : ComponentBase
 {
-    [Inject] ICategoryService CategoryService { get; set; }
-    [Parameter] public Modal Modal { get; set; }
-    [Parameter] public Category Category { get; set; }
+    [Inject] ICategoryService? CategoryService { get; set; }
+    [Parameter] public Modal? Modal { get; set; }
+    [Parameter] public Category? Category { get; set; }
     [Parameter] public EventCallback<Category> OnCategoryCreated { get; set; }
     
     private async Task HandleValidSubmit()
     {
+        if (Category == null || CategoryService == null) return;
+        
         await CategoryService.DeleteCategory(Category.Id);
-        await Hide();
         await OnCategoryCreated.InvokeAsync(Category);
+        await Hide();
     }
 
     private async Task Hide()
     {
+        if (Modal == null) return;
+        
         await Modal.HideAsync();
     }
 }
